requirements.txt
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
schedule==1.0.0
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
